Fixed organizer page not updating bug
